Renaming file casing in git
