+++ packages/ui/__tests__/sub-components/no-results-found-test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { render } from 'ink-testing-library';
 import stripAnsi from 'strip-ansi';
-import { NoResults } from '../../src/sub-components/NoResults';
+import { NoResults } from '../../src/components/NoResults';
 
 describe('Test Bar component', () => {
   it('can generate bar component', async () => {
