+++ packages/ui/src/components/List.tsx
@@ -3,7 +3,7 @@ import { Box, Text } from 'ink';
 import { RuleResults } from '@checkup/core';
 import * as objectPath from 'object-path';
 import startCase from 'lodash.startcase';
-import { TaskDisplayName } from '../sub-components/TaskDisplayName.js';
+import { TaskDisplayName } from '../components/TaskDisplayName.js';
 import { getOptions } from '../get-options.js';
 
 type ListOptions = {
