+++ packages/ui/src/components/SectionedBar.tsx
